requirements.txt
blinker>=1.5
Flask>=2.0
itsdangerous>=2.0
Jinja2>=3.0
MarkupSafe>=2.0
psycopg2-binary>=2.0
python-dotenv>=0.10
Werkzeug>=1.0
